misc/HtmlTestSuite/results/test-0528.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Grid Layout Test: Self-Alignment of absolute positioned items with 'auto' grid positions and margin, border and padding</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez Garcia-Boente"/>
<link href="https://drafts.csswg.org/css-grid-1/#column-align" rel="help"/>
<link href="https://drafts.csswg.org/css-grid-1/#abspos-items" rel="help"/>
<link href="https://drafts.csswg.org/css-grid-1/#augmented-grid" rel="help"/>
<link href="https://drafts.csswg.org/css-align-3/#propdef-justify-self" rel="help"/>
<link href="https://drafts.csswg.org/css-align-3/#propdef-align-self" rel="help"/>
<meta content="Absolute positioned grid items alignment works as expected when items have margins, borders and paddings." name="assert"/>
<link href="/fonts/ahem.css" rel="stylesheet" type="text/css"/>
<style>
.grid {
  position: relative;
  display: inline-grid;
  grid-template-columns: 100px 150px;
  grid-template-rows: 150px 100px;
  font: 10px/1 Ahem;
  background: grey;
  justify-content: center;
}
.grid.LTR, .grid.RTL { width: 450px; }
.grid.verticalLR, .grid.verticalRL { height: 450px; }
.grid &gt; div { position: absolute; }
.grid &gt; :nth-child(1) { background: green; }
.grid &gt; :nth-child(2) { background: blue; }
.grid &gt; :nth-child(3) { background: yellow; }
.grid &gt; :nth-child(4) { background: red; }

.grid &gt; div {
   margin: 2px 4px 6px 8px;
   padding: 3px 2px 4px 5px;
   border-color: silver;
   border-style: solid;
   border-width: 1px 6px 2px 3px;
}

.LTR { direction: ltr; }
.RTL { direction: rtl; }
.verticalLR { writing-mode: vertical-lr; }
.verticalRL { writing-mode: vertical-rl; }

.firstRowFirstColumn {
  grid-row: 1 / 2;
  grid-column: auto / 1;
  align-self: start;
  justify-self: right;
}
.firstRowSecondColumn {
  grid-row: 1 / 2;
  grid-column: 2 / auto;
  align-self: end;
  justify-self: center;
}
.secondRowFirstColumn {
  grid-row: 2 / 3;
  grid-column: auto / auto;
  align-self: center;
  justify-self: self-start;
}
.secondRowSecondColumn {
  grid-row: 2 / 3;
  grid-column: 1 / 2;
  align-self: self-end;
  justify-self: left;
}
</style>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<script type="text/javascript">
  setup({ explicit_done: true });
</script>
</head><body onload="document.fonts.ready.then(() =&gt; { checkLayout('.grid'); })">
<div class="grid LTR">
  <div class="firstRowFirstColumn" data-expected-height="20" data-expected-width="76" data-offset-x="20" data-offset-y="2">X XX X</div>
  <div class="firstRowSecondColumn" data-expected-height="40" data-expected-width="86" data-offset-x="284" data-offset-y="104">XX X<br/>X XXX X<br/>XX XXX</div>
  <div class="secondRowFirstColumn" data-expected-height="20" data-expected-width="76" data-offset-x="8" data-offset-y="188">X XX X</div>
  <div class="secondRowSecondColumn" data-expected-height="50" data-expected-width="76" data-offset-x="108" data-offset-y="194">XX X<br/>X XXX<br/>X<br/>XX XXX</div>
</div>

<div class="grid RTL">
  <div class="firstRowFirstColumn" data-expected-height="20" data-expected-width="76" data-offset-x="370" data-offset-y="2">X XX X</div>
  <div class="firstRowSecondColumn" data-expected-height="40" data-expected-width="86" data-offset-x="84" data-offset-y="104">XX X<br/>X XXX X<br/>XX XXX</div>
  <div class="secondRowFirstColumn" data-expected-height="20" data-expected-width="76" data-offset-x="370" data-offset-y="188">X XX X</div>
  <div class="secondRowSecondColumn" data-expected-height="50" data-expected-width="76" data-offset-x="258" data-offset-y="194">XX X<br/>X XXX<br/>X<br/>XX XXX</div>
</div>

<br/><br/>

<div class="grid verticalLR">
  <div class="firstRowFirstColumn" data-expected-height="70" data-expected-width="26" data-offset-x="8" data-offset-y="24">X XX X</div>
  <div class="firstRowSecondColumn" data-expected-height="80" data-expected-width="46" data-offset-x="100" data-offset-y="283">XX X<br/>X XXX X<br/>XX XXX</div>
  <div class="secondRowFirstColumn" data-expected-height="70" data-expected-width="26" data-offset-x="189" data-offset-y="2">X XX X</div>
  <div class="secondRowSecondColumn" data-expected-height="70" data-expected-width="56" data-offset-x="190" data-offset-y="102">XX X<br/>X XXX<br/>X<br/>XX XXX</div>
</div>

<div class="grid verticalRL">
  <div class="firstRowFirstColumn" data-expected-height="70" data-expected-width="26" data-offset-x="220" data-offset-y="24">X XX X</div>
  <div class="firstRowSecondColumn" data-expected-height="80" data-expected-width="46" data-offset-x="108" data-offset-y="283">XX X<br/>X XXX X<br/>XX XXX</div>
  <div class="secondRowFirstColumn" data-expected-height="70" data-expected-width="26" data-offset-x="39" data-offset-y="2">X XX X</div>
  <div class="secondRowSecondColumn" data-expected-height="70" data-expected-width="56" data-offset-x="8" data-offset-y="102">XX X<br/>X XXX<br/>X<br/>XX XXX</div>
</div>
</body>
</html>
